kernel bacon: replace torch gesture in favour of silent/vibrate/ring
This is just a naming change to be consistent. The actual change is done
in DeviceHandler.

Torch should be nothing which ppl activate by a gesture if the device is off.
If the user does not see that he or she enabled it it can damage on longrun the
led. Aside that we have double tap torch on lockscreen.

In meetings often it is needed to fast silent the phone. V gesture alters
now between silent / vibrate / sound.

Change-Id: I416f9fe970704b8d1ce54335f8a5134813d64607